requirements.txt
sounddevice==0.4.6
vosk==0.3.45
google-generativeai==0.3.2
python-dotenv==1.0.0
numpy==1.24.3 